When comparing timezones, explicitly use UTC.
If not setting a timezone, the timezone will be timezone naive. However, this will get serialized to a timezone-aware
timezone when being stored as a session object. This causes later comparisons to fail, since
we end up compariong timezone aware to timezone naive objects. Instead, always explicitly create timezone-aware datetimes