Use a pure path rather than real path when defining config path (#665)
* fix: use a pure path rather than real path when defining config path in templates

real paths will throw if they are instantiated on the wrong system, which causes protable builds made in windows to fail when deploying in
linux containers, e.g. on radix

* Revert "fix: use a pure path rather than real path when defining config path in templates"

This reverts commit b9d68f92749a81fa642817c9b5b89814ee48cfde.
The intended fix is not adequate, since pure paths are considered the
parent clases of real paths, the fix causes type checking to fail for
plugins

* Special case windows-path and PosixPath and replace them by the string Path when generating templates

WindowsPaths and PosixPaths are not portable, as they can not be instantiated on the 'wrong' systems. Work around this by hand-replacing the
string representation with 'Path', as this both allows system instantiation _and_ passes signature checks